Move doc tests
